requirements.txt
fastapi
uvicorn
pillow
moviepy
pyttsx3
